+++ src/components/TeacherComponents/CopyrightRestrictionBanner/CopyrightRestrictionBanner.tsx
@@ -85,12 +85,10 @@ const SignedOutCopyrightBanner = ({
           {showOnboardingLink ? (
             <OakSecondaryLink
               data-testid="copyright-banner-onboarding-link"
-              onClick={() =>
-                router.push({
-                  pathname: resolveOakHref({ page: "onboarding" }),
-                  query: { returnTo: router.asPath },
-                })
-              }
+              href={resolveOakHref({
+                page: "onboarding",
+                query: { returnTo: router.asPath },
+              })}
             >
               sign in.
             </OakSecondaryLink>
@@ -98,7 +96,10 @@ const SignedOutCopyrightBanner = ({
             <SignUpButton
               forceRedirectUrl={`/onboarding?returnTo=${router.asPath}`}
             >
-              <OakSecondaryLink data-testid="copyright-banner-signin-link">
+              <OakSecondaryLink
+                data-testid="copyright-banner-signin-link"
+                element="button"
+              >
                 sign in.
               </OakSecondaryLink>
             </SignUpButton>